% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{data_generation}
\alias{data_generation}
\title{Simulation data generation}
\usage{
data_generation(nu0, var0, k0, mu0, n, n1, n2, initProb, transProb)
}
\arguments{
\item{nu0}{Hyperparameter nu0.}

\item{var0}{Hyperparameter var0.}

\item{k0}{Hyperparameter k0.}

\item{mu0}{Hyperparameter mu0.}

\item{n1}{Hyperparameter n1.}

\item{n2}{Hyperparameter n2.}

\item{initProb}{Initial state probabilities.}

\item{transProb}{Transition probability matrix.}
}
\value{
The generated data and the true hidden states.
}
\description{
Generate data for simulations
}
