% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissionDensity.R
\name{emission_probs}
\alias{emission_probs}
\title{Emission probabilities of the data}
\usage{
emission_probs(nu, var, k, mu, A, n1, n2)
}
\arguments{
\item{nu}{The hyperparameter nu.}

\item{var}{The hyperparameter var.}

\item{k}{The hyperparameter k.}

\item{mu}{The hyperparameter mu.}

\item{A}{The input data.}

\item{n1}{The sample size of group 1.}

\item{n2}{The sample size of group 2.}
}
\value{
The data frame that has emission densities of 4 hidden states as columns and test sites as rows.
}
\description{
Generate the emission densities of 4 hidden states for all test sites in the proposed hierarchical mixture model setting.
}
