% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initEstimate.R
\name{init_est}
\alias{init_est}
\title{Initial parameter estimate}
\usage{
init_est(dat_df, mean_thresholdPV, var_thresholdPV, n1, n2, n)
}
\arguments{
\item{dat_df}{The input data frame with case 3 removed.}

\item{mean_thresholdPV}{The p-value threshold for mean test to subset case 1 data.}

\item{var_thresholdPV}{The p-value threshold for variance test to subset case 1 data.}

\item{n1}{The sample size of group 1.}

\item{n2}{The sample size of group 2.}

\item{n}{The total number of tests (test sites, methylation probes).}
}
\value{
Initial parameter estimate of state proportion parameters and hyper-parameters.
}
\description{
Subset out the case 3 data with same mean and different variance. The case 3 data
}
