% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHMMinfer.R
\name{posterior_inference}
\alias{posterior_inference}
\title{Posterior inference based on HMM estimates.}
\usage{
posterior_inference(alpha, beta, Z = NULL, train = TRUE)
}
\arguments{
\item{alpha}{The forward probabilities.}

\item{beta}{The backward probabilities.}

\item{Z}{The true hidden states.}

\item{train}{Default: True. Training mode with known true hidden states.}
}
\value{
A list of state inference and accuracy measure if in training setting. State inference if in testing setting (i.e. true hidden states are unknown).
}
\description{
Make posterior inference based on HMM estimates.
}
