% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initEstimate.R
\name{remove_case3}
\alias{remove_case3}
\title{Subset out the case 3 data}
\usage{
remove_case3(dat, mean_thresholdPV, var_thresholdPV, n1, n2)
}
\arguments{
\item{dat}{The input data (matrix) with rows as different test sites, columns as the samples from two groups with sample size as n1, n2.}

\item{mean_thresholdPV}{The p-value threshold for mean test to subset out case 3 data.}

\item{var_thresholdPV}{The p-value threshold for variance test to subset out case 3 data.}

\item{n1}{The sample size of group 1.}

\item{n2}{The sample size of group 2.}
}
\value{
The data frame that has case 3 removed.
}
\description{
Subset out the case 3 data with same mean and different variance. The case 3 data
}
