% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEMindep.R
\name{runEM_1}
\alias{runEM_1}
\title{EM on the data without case 3.}
\usage{
runEM_1(dat_df, EM_threshold = 1e-04, n1, n2, init_para_est, niter)
}
\arguments{
\item{dat_df}{The input data frame.}

\item{EM_threshold}{The threshold of parameter diffrences among iterations. Default: 0.0001. Control the number of EM iterations.}

\item{n1}{The sample size of group 1.}

\item{n2}{The sample size of group 2.}

\item{init_para_est}{The initial parameter estimates.}

\item{niter}{The upper limit for number of EM iterations.}
}
\value{
The number of iterations and the EM estimate of hyperparameters together with proportion parameters in the setting of no case 3.
}
\description{
Simply take the data without case 3 as independent Gaussian mixture and estimate the hyperparameters.
}
