% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHMMinfer.R
\name{runHMM}
\alias{runHMM}
\title{Forward-backward for HMM estimates.}
\usage{
runHMM(emissions)
}
\arguments{
\item{emissions}{The input data frame of emission densities for different test sites and 4 hidden states.}
}
\value{
The list of forward probabilites, backward probabilities, the number of forward-backward iterations, estimate of initial state probabilities and transition probabilities.
}
\description{
Run forward-backward for the HMM on the data with hyperparameters fixed (emission densities fixed).
}
