% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{truePara_generation}
\alias{truePara_generation}
\title{Simulation parameter generation}
\usage{
truePara_generation(n_state)
}
\arguments{
\item{n_state}{Default: 4. Number of hidden states}
}
\value{
The list of true parameters for simulations.
}
\description{
Generate true parameters for simulations
}
